abbreviation.md: Add User Mode Linux (UML)
The arch/x86/um directory in the Linux kernel source code contains the
architecture-specific code for the x86 architecture when running the kernel in
User Mode Linux (UML). UML is a port of the Linux kernel that allows it to run
as a user-space application, which is useful for development, testing, and
debugging purposes.

Signed-off-by: Rong Tao <rongtao@cestc.cn>