oom: oom_helpers.c: add str2size(), oom.c support KB,MB,GB suffix
process 1 be killed, Here we need to make an assumption. If it is a redis
node, redis occupies a lot of memory as cache. At this time, due to an
abnormal process, a large amount of memory is allocated in a short period
of time, which will cause redis to be killed. This is not what we expect.
The process that allocates memory the fastest should be killed. Can this
be written into a patent?

1.
    $ ./oom -v -s 3GB --oom_score_adj -1000
    set oom_score_adj to -1000
    get oom_adj 0
    get oom_score_adj 0
    get oom_score 666
    Hold 3221225472 B (3072MiB) of memory
    Killed

2.
    $ ./oom -v
    get oom_adj 0
    get oom_score_adj 0
    get oom_score 666
    OOMing...
    ^Clocated 5076172800 B (4841 MiB, 4 GiB), 807.32 MiB/s, oom_score 11128

Signed-off-by: Rong Tao <rongtao@cestc.cn>