fakeCUDA: Modify cuda library map file
This is strange. CUDA's dynamic libraries don't provide multiple interface
versions. Instead, each version of the dynamic library provides its own
version. For example,

libcublas.so.13 only provides symbols for @libcublas.so.13 and not for
@libcublas.so.12. This severely impacts dynamic library compatibility and
should be optimized.

If CUDA resolves this issue, perhaps a single .map file will suffice.

Signed-off-by: Rong Tao <rongtao@cestc.cn>