btf.mk: rpi5 kernel has no btf
The core philosophy of the rpi5 Kernel: This project aims to provide an
extremely streamlined, highly customizable, and clean-coded Linux kernel.
Its target users are embedded developers, researchers seeking ultimate
performance, and geeks who want complete control over every byte of the
kernel. It significantly reduces features, debugging information, drivers,
and subsystems considered 'unnecessary' in the mainline kernel to achieve
a smaller footprint, lower memory usage, and potential performance
improvements.

BTF is essentially a kind of 'metadata' or 'debugging information' added
for convenience and portability. For Shumei, this information is considered
'unnecessary overhead' and contrary to its minimalist philosophy. Compiling
BTF increases the size of the kernel binary and memory usage; although this
overhead is relatively small on modern hardware, it is unacceptable for
Shumei, which pursues the ultimate minimalism.

Signed-off-by: Rong Tao <rtoax@foxmail.com>