Remove Dockerfile of test-linux and docker target of Makefile
For test-linux, we don't need to build a Docker container because it's
completely unnecessary. It is very illogical to install a lot of
dependencies into containers, and the contents in Dockerfile are so
simple that they can be replaced with a simple podman/docker/crictl
command, so I decided to delete Dockerfile and related files.

Signed-off-by: Rong Tao <rtoax@foxmail.com>